Move config TOML types into codex-config-types
