+++ codex-rs/config-types/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "config-types",
+    crate_name = "codex_config_types",
+)
