+++ codex-rs/core/src/bin/config_schema.rs
@@ -15,6 +15,6 @@ fn main() -> Result<()> {
     let out_path = args
         .out
         .unwrap_or_else(|| PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("config.schema.json"));
-    codex_core::config::schema::write_config_schema(&out_path)?;
+    codex_config_types::schema::write_config_schema(&out_path)?;
     Ok(())
 }
