+++ codex-rs/core/src/config/config_tests.rs
@@ -17,9 +17,32 @@ use codex_config::types::MemoriesToml;
 use codex_config::types::ModelAvailabilityNuxConfig;
 use codex_config::types::NotificationMethod;
 use codex_config::types::Notifications;
+use codex_config::types::SandboxWorkspaceWrite;
+use codex_config::types::SkillsConfig;
 use codex_config::types::ToolSuggestDiscoverableType;
+use codex_config::types::Tui;
+use codex_config_types::AgentRoleToml;
+use codex_config_types::AgentsToml;
+use codex_config_types::ConfigProfile;
+use codex_config_types::ConfigToml;
+use codex_config_types::FilesystemPermissionToml;
+use codex_config_types::FilesystemPermissionsToml;
+use codex_config_types::NetworkDomainPermissionToml;
+use codex_config_types::NetworkDomainPermissionsToml;
+use codex_config_types::NetworkToml;
+use codex_config_types::PermissionProfileToml;
+use codex_config_types::PermissionsToml;
+use codex_config_types::ProjectConfig;
+use codex_config_types::RealtimeAudioConfig;
+use codex_config_types::RealtimeConfig;
+use codex_config_types::RealtimeToml;
+use codex_config_types::RealtimeWsMode;
+use codex_config_types::RealtimeWsVersion;
+use codex_config_types::ToolsToml;
 use codex_features::Feature;
 use codex_features::FeaturesToml;
+use codex_model_provider_info::LMSTUDIO_OSS_PROVIDER_ID;
+use codex_model_provider_info::OLLAMA_OSS_PROVIDER_ID;
 use codex_model_provider_info::WireApi;
 use codex_models_manager::bundled_models_response;
 use codex_protocol::permissions::FileSystemAccessMode;
@@ -28,6 +51,7 @@ use codex_protocol::permissions::FileSystemSandboxEntry;
 use codex_protocol::permissions::FileSystemSandboxPolicy;
 use codex_protocol::permissions::FileSystemSpecialPath;
 use codex_protocol::permissions::NetworkSandboxPolicy;
+use codex_protocol::protocol::ReadOnlyAccess;
 use serde::Deserialize;
 use tempfile::tempdir;
