+++ codex-rs/core/src/config/managed_features.rs
@@ -8,8 +8,8 @@ use codex_config::FeatureRequirementsToml;
 use codex_config::RequirementSource;
 use codex_config::Sourced;
 
-use crate::config::ConfigToml;
-use crate::config::profile::ConfigProfile;
+use codex_config_types::ConfigProfile;
+use codex_config_types::ConfigToml;
 use codex_features::Feature;
 use codex_features::FeatureConfigSource;
 use codex_features::FeatureOverrides;
