+++ codex-rs/core/tests/suite/live_reload.rs
@@ -6,7 +6,7 @@ use std::path::PathBuf;
 use std::time::Duration;
 
 use anyhow::Result;
-use codex_core::config::ProjectConfig;
+use codex_config_types::ProjectConfig;
 use codex_protocol::config_types::TrustLevel;
 use codex_protocol::protocol::AskForApproval;
 use codex_protocol::protocol::EventMsg;
